containers/hailo-ai-service/requirements.txt
# Hailo8 AI Service Dependencies

# Web framework
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# HTTP client
requests==2.31.0
aiohttp==3.9.1

# Image processing
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# Data processing
pandas==2.1.4
matplotlib==3.8.2

# Async support
asyncio==3.4.3

# Logging
structlog==23.2.0

# Configuration
python-dotenv==1.0.0

# Image and video processing
imageio==2.31.6
imageio-ffmpeg==0.4.9

# Machine learning utilities
scikit-learn==1.3.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Performance monitoring
psutil==5.9.6

# File handling
python-multipart==0.0.6